Rename `rustc_hir_monomorphize/src/errors.rs` into `rustc_hir_monomorphize/src/diagnostics.rs`
