+++ compiler/rustc_monomorphize/src/lib.rs
@@ -13,7 +13,7 @@ use rustc_middle::{bug, traits};
 use rustc_span::ErrorGuaranteed;
 
 mod collector;
-mod errors;
+mod diagnostics;
 mod graph_checks;
 mod mono_checks;
 mod partitioning;
